Add new js file for Clear All Completed function
