+++ src/index.html
@@ -7,6 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <script src="https://kit.fontawesome.com/773702659e.js" crossorigin="anonymous"></script>
     <title>To Do List</title>
+    <link rel="stylesheet" href="style.css">
 </head>
 
 <body>
@@ -14,10 +15,11 @@
         <h2 id="title">Today's To Do</h2>
         <input type="text" id="input-field" placeholder="Add to your list..." required>
         <div id="list-container">
-            <!-- List using JavaScript -->
+            <!-- List Generated using JavaScript -->
         </div>
         <h2 id="clear-button">Clear all completed</h2>
     </section>
+    <script type="module" src="./index.js"></script>
 </body>
 
 </html>
