+++ src/index.js
@@ -1,6 +1,8 @@
 import ToDoList from './toDoList.js';
+import RemoveCheckbox from './removeCheckbox.js';
 
 const toDoList = new ToDoList();
+const removeCheckboxObject = new RemoveCheckbox(toDoList);
 
 document.addEventListener('keyup', (event) => {
   const inputField = document.getElementById('input-field');
@@ -12,7 +14,6 @@ document.addEventListener('keyup', (event) => {
 });
 
 const clearButton = document.getElementById('clear-button');
-
 clearButton.addEventListener('click', () => {
-  toDoList.removeCompletedTasks();
+  removeCheckboxObject.removeCompletedTasks();
 });
